+++ session4/hw4/GrowArray/src/Point.java
@@ -1,7 +1,7 @@
 public class Point {
-    int x, y;
+    double x, y;
 
-    public Point(int x, int y) {
+    public Point(double x, double y) {
         this.x = x;
         this.y = y;
     }
